requirements.txt
streamlit==1.36.0
langchain==0.2.11
langchain-core==0.2.29
langchain-groq==0.1.9
python-dotenv==1.0.1
langdetect==1.0.9